# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FreightAudit API — análise, auditoria e simulação de remuneração de fretes
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/recent-changes:
    get:
      operationId: getDashboardRecentChanges
      tags: [dashboard]
      summary: Alterações recentes detectadas
      parameters:
        - name: limit
          in: query
          schema:
            type: number
            default: 10
      responses:
        "200":
          description: Lista de alterações recentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiffItem"

  # ─── SNAPSHOTS ───────────────────────────────────────────────────────────────
...
  /diffs:
    get:
      operationId: listDiffs
      tags: [diffs]
      summary: Listar comparações entre snapshots
      responses:
        "200":
          description: Lista de diffs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SnapshotDiff"
    post:
      operationId: createDiff
      tags: [diffs]
      summary: Gerar comparação entre dois snapshots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiffInput"
      responses:
        "201":
          description: Diff gerado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotDiff"
        "400":
          description: Parâmetros inválidos
  /diffs/{id}:
    get:
      operationId: getDiff
      tags: [diffs]
      summary: Obter diff por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Diff encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotDiff"
        "404":
          description: Diff não encontrado
  /diffs/{id}/items:
    get:
      operationId: getDiffItems
      tags: [diffs]
      summary: Listar itens de alteração de um diff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: changeType
          in: query
          schema:
            type: string
        - name: changeNature
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Itens do diff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiffItem"

  # ─── IMPORTS ─────────────────────────────────────────────────────────────────
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: Listar alertas do sistema
      parameters:
        - name: isRead
          in: query
          schema:
            type: boolean
        - name: severity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de alertas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Marcar alerta como lido ou dispensado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertUpdate"
      responses:
        "200":
          description: Alerta atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Alerta não encontrado

  # ─── SIMULATIONS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SnapshotDiff:
      type: object
      required:
        - id
        - snapshotAId
        - snapshotBId
        - snapshotALabel
        - snapshotBLabel
        - status
        - totalAdded
        - totalRemoved
        - totalChanged
        - totalUnchanged
        - generatedAt
      properties:
        id:
          type: string
        snapshotAId:
          type: string
        snapshotBId:
          type: string
        snapshotALabel:
          type: string
        snapshotBLabel:
          type: string
        status:
          type: string
          enum: [PENDING, DONE, FAILED]
        totalAdded:
          type: number
        totalRemoved:
          type: number
        totalChanged:
          type: number
        totalUnchanged:
          type: number
        estimatedMonthlyImpact:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        generatedAt:
          type: string
          format: date-time
    DiffInput:
      type: object
      required: [snapshotAId, snapshotBId]
      properties:
        snapshotAId:
          type: string
        snapshotBId:
          type: string
        notes:
          type: string
    DiffItem:
      type: object
      required:
        - id
        - diffId
        - entityType
        - entityKey
        - entityName
        - changeType
      properties:
        id:
          type: string
        diffId:
          type: string
        entityType:
          type: string
          enum: [PARAMETER, TARIFF_TABLE, RULE]
        entityKey:
          type: string
        entityName:
          type: string
        changeType:
          type: string
          enum: [ADDED, REMOVED, CHANGED, UNCHANGED]
        changeNature:
          type: ["string", "null"]
        severity:
          type: ["string", "null"]
          enum: [LOW, MEDIUM, HIGH, CRITICAL, UNKNOWN, null]
        valueBefore:
          type: ["number", "null"]
        valueAfter:
          type: ["number", "null"]
        valueBeforeText:
          type: ["string", "null"]
        valueAfterText:
          type: ["string", "null"]
        deltaAbsolute:
          type: ["number", "null"]
        deltaPercent:
          type: ["number", "null"]
        estimatedMonthlyImpact:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        snapshotALabel:
          type: ["string", "null"]
        snapshotBLabel:
          type: ["string", "null"]

    # ─── IMPORT JOB ─────────────────────────────────────────────────────────────
...
    Alert:
      type: object
      required: [id, alertType, severity, title, isRead, isDismissed, createdAt]
      properties:
        id:
          type: string
        alertType:
          type: string
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        title:
          type: string
        body:
          type: ["string", "null"]
        estimatedImpact:
          type: ["number", "null"]
        isRead:
          type: boolean
        isDismissed:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AlertUpdate:
      type: object
      properties:
        isRead:
          type: boolean
        isDismissed:
          type: boolean

    # ─── SIMULATION ─────────────────────────────────────────────────────────────
...